back/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dictdiffer==0.9.0
ecdsa==0.19.0
et-xmlfile==1.1.0
fastapi==0.112.0
h11==0.14.0
idna==3.7
iso8601==1.1.0
kombu==5.4.0
openpyxl==3.1.5
prompt_toolkit==3.0.47
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
redis==5.0.8
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomlkit==0.13.1
tortoise-orm==0.21.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
